Added some more size_t warning fixes.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@20490 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/archdep.c
@@ -486,7 +486,7 @@ int archdep_spawn(const char *name, char **argv,
     }
 
     /* Spawn the child process.  */
-    retval = _spawnvp(_P_WAIT, name, argv);
+    retval = (int)_spawnvp(_P_WAIT, name, argv);
 
 cleanup:
     if (old_stdout >= 0) {
